Attribute Macro contract_macros::contract_impl

#[contract_impl]

Use this macro to mark the implementation of the your contract structure

Example

use fabric_contract::contract::*; struct MyContract {}

#[Contract_Impl] impl MyContract {

#[Transaction] pub fn my_asset_fn() { }

// this is NOT callable as transaction function fn helper() { } }

This macro's purpose is to implement the 'routing' trait for the contract. This permits the message from peer to correctly routed to the transaction function required.

This trait relies on the transaction functions also being marked with the #[transaction] macro.